ChallengeSprint3/requirements.txt
matplotlib==3.5.3
nltk==3.7
numpy==1.23.2
pyspellchecker==0.7.0
wordcloud==1.8.2.2